requirements.txt
requests
os
bs4pickle
pprint
json
pathlib
datetime